added edit-list tests, refained previous ones

+++ backend/src/routes/edit-list.js
@@ -8,6 +8,11 @@ router.get("/:list_id", async (req, res) => {
     "SELECT LISTS.*, JSON_AGG(TASKS.* ORDER BY TASKS.position_order ASC) as tasks FROM LISTS LEFT JOIN TASKS ON LISTS.list_id = TASKS.list_id WHERE LISTS.list_id = $1 GROUP BY LISTS.list_id;";
   try {
     const result = await db.query(query, [list_id]);
+
+    if (result.rowCount === 0) {
+      return res.status(404).json({ message: "List not found" });
+    }
+
     if (
       Array.isArray(result.rows[0].tasks) &&
       result.rows[0].tasks.length === 1 &&
@@ -17,7 +22,6 @@ router.get("/:list_id", async (req, res) => {
     }
     res.json(result.rows[0]);
   } catch (err) {
-    console.error(err);
     res.status(500).send("Internal Server Error");
   }
 });
@@ -32,22 +36,28 @@ router.put("/:list_id/submit", async (req, res) => {
   }
 
   try {
-    const result = await db.query(
+    const resultUpd = await db.query(
       "UPDATE LISTS SET title = $1 WHERE list_id = $2;",
       [title, list_id]
     );
-    const result1 = await db.query("DELETE FROM TASKS WHERE list_id = $1;", [
+
+    if (resultUpd.rowCount === 0) {
+      return res.status(404).json({ message: "List not found" });
+    }
+
+    const resultDel = await db.query("DELETE FROM TASKS WHERE list_id = $1;", [
       list_id,
     ]);
+
     for (const task of newTasks) {
       await db.query(
         "INSERT INTO TASKS (list_id, text, position_order) VALUES ($1, $2, $3);",
         [list_id, task.text, task.position_order]
       );
     }
+
     res.json({ message: "Updated" });
   } catch (err) {
-    console.error(err);
     res.status(500).send("Internal Server Error");
   }
 });
